+++ sdk/nodejs/route53/resolverDnsSecConfig.ts
@@ -62,10 +62,6 @@ export class ResolverDnsSecConfig extends pulumi.CustomResource {
      */
     public /*out*/ readonly arn!: pulumi.Output<string>;
     /**
-     * The ID for a configuration for DNSSEC validation.
-     */
-    public /*out*/ readonly id!: pulumi.Output<string>;
-    /**
      * The owner account ID of the virtual private cloud (VPC) for a configuration for DNSSEC validation.
      */
     public /*out*/ readonly ownerId!: pulumi.Output<string>;
@@ -92,7 +88,6 @@ export class ResolverDnsSecConfig extends pulumi.CustomResource {
         if (opts.id) {
             const state = argsOrState as ResolverDnsSecConfigState | undefined;
             inputs["arn"] = state ? state.arn : undefined;
-            inputs["id"] = state ? state.id : undefined;
             inputs["ownerId"] = state ? state.ownerId : undefined;
             inputs["resourceId"] = state ? state.resourceId : undefined;
             inputs["validationStatus"] = state ? state.validationStatus : undefined;
@@ -103,7 +98,6 @@ export class ResolverDnsSecConfig extends pulumi.CustomResource {
             }
             inputs["resourceId"] = args ? args.resourceId : undefined;
             inputs["arn"] = undefined /*out*/;
-            inputs["id"] = undefined /*out*/;
             inputs["ownerId"] = undefined /*out*/;
             inputs["validationStatus"] = undefined /*out*/;
         }
@@ -122,10 +116,6 @@ export interface ResolverDnsSecConfigState {
      * The ARN for a configuration for DNSSEC validation.
      */
     arn?: pulumi.Input<string>;
-    /**
-     * The ID for a configuration for DNSSEC validation.
-     */
-    id?: pulumi.Input<string>;
     /**
      * The owner account ID of the virtual private cloud (VPC) for a configuration for DNSSEC validation.
      */